DEV: Fix JS acceptance test (#8)
I noticed this flaky test when debugging the plugins tests failures in core PR https://github.com/discourse/discourse/pull/11726, but this is unrelated to changes in that PR.

The topic timeline controls on the right-hand side are hidden if the window width is below a certain threshold. So, if you don't make your window wide enough when running tests, the test fails because the reply button of the timeline is gone.

This PR changes the test so it uses the reply button at the bottom of the topic instead of the reply button of the timeline controls.