FIX: Allow wiki and post_type revisions on staff alias user post.
Previously, a user that is allowed to post as a staff alias user would
not be able to change the post type or wiki status of the post. This
commit fixes that by automatically revising the post as the staff alias
user if the user making the revision is allowed to post as a staff alias
user.