FIX: Allow any type of post revision for staff alias topics (#86)
Currently, when a post is modified, and the post is created by the "staff alias user", we go through some validations like

is the user trying to edit a whisper? (not allowed)
is the post modification done by "a user who can use the staff alias (SiteSetting.staff_alias_allowed_groups)? (ok)
the modification is a wiki, post_type, user_id, title, tag -type modification.
When a new modification-type is missing, the user will face a 422 error.

This PR removes the criteria 3 to allow any modifications as long as the acting user is part of criteria 2. Related fixes in the past (73dffcd, 8762cc2, 291999b) were merely adding on to criteria 3 and it doesn't make sense long term.